[frontend]Unify simulation challenge
